requirements.txt
Django==3.1.7
django-bootstrap4==2.3.1
django-simple-captcha==0.5.14
django-tinymce==3.3.0
psycopg2-binary==2.8.6
pillow==8.2.0
